Refactored CpuCustomNonbondedForce to support multiple vector sizes
